% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anovaOnCNV.R
\name{anovaOnCNV}
\alias{anovaOnCNV}
\title{anovaOnCNV}
\usage{
anovaOnCNV(RE, cnv, anovacov, nameofnewdf = "CNVanova")
}
\arguments{
\item{RE}{RaggedExperiment object used to store all information.}

\item{cnv}{List of CNVs in a dataframe containing CNVs from detection algorithms/ pipelines.}

\item{anovacov}{List of dataframes which contain the required information to perform anova for each region with a detected CNV. Generated from prepAnova, and stored as metadata.}

\item{nameofnewdf}{Name of new dataframe to be saved in metadata(RE)[['CNVlist']]. Default is CNVanova.}
}
\value{
A new cnv dataframe with an additional column for the p-value from anova. Will be stored as an experiment in the RE object.
}
\description{
Apply anova on the CNV data. For each CNV listed in the input cnv file, the read-depths will be used to calculate if this region of the DNA, in this sample, is significantly different from other samples in the cohort of samples, at this same region.
}
\examples{
data("test_cnv")
data("test_cov")
SARC <- regionSet(cnv = test_cnv, cov = test_cov)
SARC <- regionSplit(RE = SARC, cnv = metadata(SARC)[['CNVlist']][[1]],
                    startlist = metadata(SARC)[[2]],
                    endlist = metadata(SARC)[[3]])
SARC <- prepAnova(RE = SARC, cnv = metadata(SARC)[['CNVlist']][[1]],
                 start = metadata(SARC)[[2]], end = metadata(SARC)[[3]])
SARC <- anovaOnCNV(RE = SARC, cnv = metadata(SARC)[['CNVlist']][[1]],
                   anovacov = metadata(SARC)[[4]])
}
