% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCNV.R
\name{plotCNV}
\alias{plotCNV}
\title{plotCNV}
\usage{
plotCNV(
  bed,
  setup,
  FilteredCNV = 1,
  batch = NULL,
  nSamples = NULL,
  gene = NULL,
  log = NULL
)
}
\arguments{
\item{bed}{CNV bed files stored as dataframes. Chromosomes, Start of CNV, End of CNV, Type of CNV and some other information can be kept here. It is recomended to use the most recently generated bed file. Check print(MA) to see if bed files from other anlyses have been generated.}

\item{setup}{ist of dataframes which have been processed for plotting. This will be stored as metadata in the MA object after SARC::setupCNVPlot.}

\item{FilteredCNV}{Which CNV from the list of dataframes (setup) should be plotted. Can be automated within a for-loop. Default is 1.}

\item{batch}{If WES/ WGS was performed in batches, users may wish to plot samples from different batches separately. In this case, a column called 'BATCH' should be found in the BED file. Different batches should be labelled with differentiating strings, and this string should be added to the batch parameter. Default is NULL.}

\item{nSamples}{Numbers of samples to be printed in the plots. Default is 10. Add an integer to represent how many subplots should be made. The higher the number, the more difficult the plot will be for viewing.}

\item{gene}{String of a gene of interest. Only this gene will be printed. It must match standard gene symbols e.g. FCGR3A, P53. Only one string should be inputted at a time. The default is null.}

\item{log}{Applies log normalisation to reads. This can be helpful in WES data, where reads can be quite disperse across a short region of DNA. Reccomended is 10, and default is NULL.}
}
\value{
A grid plot showing the read-depths for a specific regions of the the genome. This region will contain one samples which will have had a CNV detected. Heat-plot like colouring will be used to visually show if the sample has a significant change at this region, in contrast to several other samples.
}
\description{
Plots of the region of the DNA from several samples where a CNV is detected. The sample with the detected CNV, from the bed file, will be highlighted in purple. Additional specificity such as gene of interest, or batch of WES/ WGS can be specified. This function is made to be easily looped for multiple CNVs. It is also made as a quick and more visually preferable alternative to looking for true CNVs visually.
}
\examples{
data("test_bed")
data("test_cov")
test_bed <- test_bed[c(1),]
SARC <- regionSet(bed = test_bed, cov = test_cov)
SARC <- plotCovPrep(MA = SARC, bed = test_bed, cov = test_cov,
                   startlist = metadata(SARC)[[1]],
                   endlist = metadata(SARC)[[2]])
SARC <- regionGrangeMake(MA = SARC, covprepped = metadata(SARC)[[3]])

library(TxDb.Hsapiens.UCSC.hg38.knownGene)
library(Homo.sapiens)
TxDb(Homo.sapiens) <- TxDb.Hsapiens.UCSC.hg38.knownGene
txdb <- TxDb.Hsapiens.UCSC.hg38.knownGene
tx <- transcriptsBy(Homo.sapiens, columns = "SYMBOL")
txgene <- tx@unlistData

SARC <- addExonsGenes(MA = SARC, covgranges = metadata(SARC)[[4]],
                      txdb = txdb, txgene = txgene)

SARC <- setupCNVplot(MA = SARC, namedgranges =  metadata(SARC)[[5]],
                  covprepped = metadata(SARC)[[3]])
p <- plotCNV(bed = test_bed, setup = metadata(SARC)[[6]], FilteredCNV=1)
}
