% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCovPrep.R
\name{plotCovPrep}
\alias{plotCovPrep}
\title{plotCovPrep}
\usage{
plotCovPrep(RE, cnv, n1 = 0, n2 = 0, startlist, endlist)
}
\arguments{
\item{RE}{RaggedExperiment object used to store all information.}

\item{cnv}{List of CNVs in a dataframe containing CNVs from detection algorithms/ pipelines. Must use one which contains at least MeanScore, Qlow, Qhigh and anova p-values. If Post-hoc tests have also been performed, these results can also be used by this function.}

\item{n1}{How many rows to pad the start site of each CNV - default is 0.}

\item{n2}{How many rows to pad the end site of each CNV - default is 0.}

\item{startlist}{List of start sites created from SARC::regionSet and stored as metadata.}

\item{endlist}{List of end sites created from SARC::regionSet and stored as metadata.}
}
\value{
A new list of dataframes, each df is a region where a CNV was detected.
}
\description{
Prepares plotting of coverage at specific regions of the DNA - where CNVs have been detected.
}
\examples{
data("test_cnv")
test_cnv <- test_cnv[c(1:3),]
data("test_cov")
SARC <- regionSet(cnv = test_cnv, cov = test_cov)
SARC <- regionSplit(RE = SARC, cnv = metadata(SARC)[['CNVlist']][[1]],
                    startlist = metadata(SARC)[[2]],
                    endlist = metadata(SARC)[[3]])
SARC <- plotCovPrep(RE = SARC, cnv = metadata(SARC)[['CNVlist']][[1]],
                    startlist = metadata(SARC)[[2]],
                    endlist = metadata(SARC)[[3]])
}
