% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepAnova.R
\name{prepAnova}
\alias{prepAnova}
\title{prepAnova}
\usage{
prepAnova(MA, bed, cov, startlist, endlist)
}
\arguments{
\item{MA}{MultiAssayExperiment object used to store all information.}

\item{bed}{Bed file containing CNVs which the user wishes to generate plots for. It is recommended that the most recently created bed file is used. Check print(MA) to see more bed files created by SARC.}

\item{cov}{cov file from WES platform/ sequencer kit or if WGS regular intervals. Stored as a dataframe - genomic locations as rows and samples as columns.}

\item{startlist}{List of start sites created from SARC::regionSet and stored as metadata.}

\item{endlist}{List of end sites created from SARC::regionSet and stored as metadata.}
}
\value{
Additional list of dataframes which will be needed to calculate anova for each region with a detected CNV.
}
\description{
Function for annovaOnCNV.This function sets up samples for anova analysis. Can also be used as input for dunnet analysis.
}
\examples{
data("test_bed")
data("test_cov")
SARC <- regionSet(bed = test_bed, cov = test_cov)
SARC <- regionSplit(MA = SARC, bed = test_bed, cov = test_cov,
                    startlist = metadata(SARC)[[1]],
                    endlist = metadata(SARC)[[2]])
SARC <- regionMean(MA = SARC, bed = test_bed, splitcov = metadata(SARC)[[3]])
SARC <- regionQuantiles(MA = SARC, bed = experiments(SARC)[[1]],
                       meancov = metadata(SARC)[[4]], q1=.15, q2=.85)
SARC <- prepAnova(MA = SARC, bed = experiments(SARC)[[2]], cov = test_cov,
                 startlist = metadata(SARC)[[1]], endlist = metadata(SARC)[[2]])
}
