% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionQuantiles.R
\name{regionQuantiles}
\alias{regionQuantiles}
\title{regionQuantiles}
\usage{
regionQuantiles(
  RE,
  cnv,
  meancov,
  q1 = 0.05,
  q2 = 0.95,
  nameofnewdf = "CNVquantiles"
)
}
\arguments{
\item{RE}{RaggedExperiment object used to store all information.}

\item{cnv}{List of CNVs in a dataframe containing CNVs from detection algorithms/ pipelines. It is recommended to use the most recently generated cnv file. Check print(RE) to see if cnv files from other analyses have been generated.}

\item{meancov}{Small cov files with meanscores included. This will be stored as metadata after being generated by SARC::regionMean.}

\item{q1}{Lower quantile cut-off. Default is .05. This should be adjusted based on the number of samples, i.e. less samples > higher the lower threshold.}

\item{q2}{Upper quantile cut-off. Default is .95. This should be adjusted based on the number of samples, i.e. less samples > lower the higher threshold.}

\item{nameofnewdf}{Name of new dataframe to be saved in metadata(RE)[['CNVlist']]. Default is CNVquantiles.}
}
\value{
A new cnv file with two new columns. For deletions the Qlow columns will be 0 or 1, and for duplications the Qhigh columns will be 0 or 1.
}
\description{
Calculates quantile distribution from each sample, within a region with a suspected CNV. We assume true deletions would be in the lower end of the quantile distribution and true duplications to be in the upper end of the quantile distribution. Many CNVs captured by detection pipelines will be false positives, this this is a good metric to remove obvious false positives. Lower quantile (q1) and upper quantile (q2) should be adjusted based on the number of patient samples available.
}
\examples{
data("test_cnv")
data("test_cov")
SARC <- regionSet(cnv = test_cnv, cov = test_cov)
SARC <- regionSplit(RE = SARC, cnv = metadata(SARC)[['CNVlist']][[1]],
                    startlist = metadata(SARC)[[2]],
                    endlist = metadata(SARC)[[3]])
SARC <- regionMean(RE = SARC, cnv = metadata(SARC)[['CNVlist']][[1]],
                   splitcov = metadata(SARC)[[4]])
SARC <- regionQuantiles(RE = SARC, cnv = metadata(SARC)[['CNVlist']][[2]],
                      meancov = metadata(SARC)[[5]], q1=.1, q2=.9)
}
