% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionSplit.R
\name{regionSplit}
\alias{regionSplit}
\title{regionSplit}
\usage{
regionSplit(RE, cnv, startlist, endlist)
}
\arguments{
\item{RE}{RaggedExperiment object used to store all information.}

\item{cnv}{List of CNVs in a dataframe containing CNVs from detection algorithms/ pipelines. It is recommended that the most recently created cnv file is used. Check print(RE) to see more cnv files created by SARC.}

\item{startlist}{List of start sites created from SARC::regionSet and stored as metadata.}

\item{endlist}{List of end sites created from SARC::regionSet and stored as metadata.}
}
\value{
A list of dataframes - each one a smaller cov file.
}
\description{
Splits the large cov file into multiple smaller cov files. Each one is specific for each CNV entry in the bam file.
}
\examples{
data("test_cnv")
data("test_cov")
SARC <- regionSet(cnv = test_cnv, cov = test_cov)
SARC <- regionSplit(RE = SARC, cnv =  metadata(SARC)[['CNVlist']][[1]],
                    startlist = metadata(SARC)[[2]],
                    endlist = metadata(SARC)[[3]])
}
