% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setDPlot.R
\name{setQDplot}
\alias{setQDplot}
\title{setDPlot}
\usage{
setQDplot(MA, meancov)
}
\arguments{
\item{MA}{MultiAssayExperiment object used to store all information.}

\item{meancov}{Small cov files with mean scores included. This will be stored as metadata after being generated by SARC::regionMean.}
}
\value{
A list of dataframes. Each dataframe contains ranked meanscores for each sample. One dataframe is made for each CNV in the bed file.
}
\description{
Quality checking plot to see the Distribution of mean scores at a given region of the genome. The sample with the suspected CNV is highlighted. Useful way of checking why a certain CNV may be a false positive.
}
\examples{
data("test_bed")
data("test_cov")
SARC <- regionSet(bed = test_bed, cov = test_cov)
SARC <- regionSplit(MA = SARC, bed = test_bed, cov = test_cov,
                    startlist = metadata(SARC)[[1]],
                    endlist = metadata(SARC)[[2]])
SARC <- regionMean(MA = SARC, bed = test_bed, splitcov = metadata(SARC)[[3]])
SARC <- regionQuantiles(MA = SARC, bed = experiments(SARC)[[1]],
                        meancov = metadata(SARC)[[3]], q1 =.1, q2 = .9)
SARC <- setQDplot(MA = SARC, meancov = metadata(SARC)[[4]])
}
