% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anovaOnCNV.R
\name{anovaOnCNV}
\alias{anovaOnCNV}
\title{anovaOnCNV}
\usage{
anovaOnCNV(MA, bed, anovacov)
}
\arguments{
\item{MA}{MultiAssayExperiment object used to store all information.}

\item{bed}{CNV bed files stored as dataframes. Chromosomes, Start of CNV, End of CNV, Type of CNV and some other information can be kept here. It is recomended to use the most recently generated bed file. Check print(MA) to see if bed files from other anlyses have been generated.}

\item{anovacov}{List of dataframes which contain the required information to perform anova for each region with a detected CNV. Generated from prepAnova, and stored as metadata.}
}
\value{
A new bed file with an additional column for the p-value from anova. Will be stored as an experiment in the MA object.
}
\description{
Apply anova on the CNV data. For each CNV listed in the input bed file, the read-depths will be used to calculate if this region of the DNA, in this sample, is significantly different from other samples in the cohort of samples, at this same region.
}
\examples{
data("test_bed")
data("test_cov")
SARC <- regionSet(bed = test_bed, cov = test_cov)
SARC <- regionSplit(MA = SARC, bed = test_bed, cov = test_cov,
                    startlist = metadata(SARC)[[1]],
                    endlist = metadata(SARC)[[2]])
SARC <- prepAnova(MA = SARC, bed = test_bed, cov = test_cov,
                 start = metadata(SARC)[[1]], end = metadata(SARC)[[2]])
SARC <- anovaOnCNV(MA = SARC, bed = test_bed,
                   anovacov = metadata(SARC)[[3]])
}
