% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvConfidence.R
\name{cnvConfidence}
\alias{cnvConfidence}
\title{cnvConfidence}
\usage{
cnvConfidence(MA, bed, ph = FALSE, m1 = 0.2, m2 = 0.8, m3 = 1.2, m4 = 1.8)
}
\arguments{
\item{MA}{MultiAssayExperiment object used to store all information.}

\item{bed}{Bed file containing CNVs for analysis. Must use one which contains at least MeanScore, Qlow, Qhigh and anova p-values. If Post-hoc tests have also been performed, these results can also be used by this function.}

\item{ph}{If Dunnet tests were also performed, the p-value from the tests can be taken into account. Default is FALSE.}

\item{m1}{Value used to cut-off the highest mean score allowed for HOMOZYGOUS DELECTIONS, and lowest cut-off for HETEROZYGOUS DELETIONS. Default is 0.2.}

\item{m2}{Value used to cut-off the highest mean score allowed for HETEROZYGOUS DELETINS. Default is 0.8.}

\item{m3}{Value used to cut-off the lowst mean score allowed for HETEROZYGOUS DUPLICATIONS. Default is 1.2.}

\item{m4}{Value used to cut-off the lowst mean score allowed for HETEROZYGOUS DUPLICATIONS. Default is 1.8.}
}
\value{
A new bed file with additional columns which describe our level of confidence of the detected CNV being a true CNV.
}
\description{
Flags the confidence of the CNV being a true CNV. The flags range from Very Unconfident - Very Confident. Our clinicians preferred variants to be flagged rather than filtered out - so we simply do this. Confidence scores will be added to the bed file given. The bed file should be one which contains all the statistical results from other functions of the SARC package.
}
\examples{
data("test_bed")
data("test_cov")
SARC <- regionSet(bed = test_bed, cov = test_cov)
SARC <- regionSplit(MA = SARC, bed = test_bed, cov = test_cov,
                     startlist = metadata(SARC)[[1]],
                      endlist = metadata(SARC)[[2]])
SARC <- regionMean(MA = SARC, bed = test_bed, splitcov = metadata(SARC)[[3]])
SARC <- regionQuantiles(MA = SARC, bed = experiments(SARC)[[1]],
                        meancov = metadata(SARC)[[3]], q1 =.1, q2 = .9)
SARC <- prepAnova(MA = SARC, bed = experiments(SARC)[[2]], cov = test_cov,
                 start = metadata(SARC)[[1]], end=metadata(SARC)[[2]])
SARC <- anovaOnCNV(MA = SARC, bed = experiments(SARC)[[2]],
                  anovacov = metadata(SARC)[[7]])
SARC <- cnvConfidence(MA = SARC, bed = experiments(SARC)[[3]])
}
