% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepAnova.R
\name{prepAnova}
\alias{prepAnova}
\title{prepAnova}
\usage{
prepAnova(RE, cnv, startlist, endlist)
}
\arguments{
\item{RE}{RaggedExperiment object used to store all information.}

\item{cnv}{List of CNVs in a dataframe containing CNVs from detection algorithms/ pipelines. It is recommended that the most recently created cnv file is used. Check print(RE) to see more cnv files created by SARC.}

\item{startlist}{List of start sites created from SARC::regionSet and stored as metadata.}

\item{endlist}{List of end sites created from SARC::regionSet and stored as metadata.}
}
\value{
Additional list of dataframes which will be needed to calculate anova for each region with a detected CNV.
}
\description{
Function for annovaOnCNV.This function sets up samples for anova analysis. Can also be used as input for Dunnet analysis.
}
\examples{
data("test_cnv")
data("test_cov")
SARC <- regionSet(cnv = test_cnv, cov = test_cov)
SARC <- regionSplit(RE = SARC, cnv = metadata(SARC)[['CNVlist']][[1]],
                    startlist = metadata(SARC)[[2]],
                    endlist = metadata(SARC)[[3]])
SARC <- regionMean(RE = SARC, cnv = metadata(SARC)[['CNVlist']][[1]],
                   splitcov = metadata(SARC)[[4]])
SARC <- regionQuantiles(RE = SARC, cnv = metadata(SARC)[['CNVlist']][[2]],
                      meancov = metadata(SARC)[[5]], q1=.1, q2=.9)
SARC <- prepAnova(RE = SARC, cnv = metadata(SARC)[['CNVlist']][[3]],
                 startlist = metadata(SARC)[[2]],
                 endlist = metadata(SARC)[[3]])
}
