% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionMean.R
\name{regionMean}
\alias{regionMean}
\title{regionMean}
\usage{
regionMean(RE, cnv, splitcov, nameofnewdf = "CNVmeans")
}
\arguments{
\item{RE}{RaggedExperiment object used to store all information.}

\item{cnv}{List of CNVs in a dataframe containing CNVs from detection algorithms/ pipelines. It is recomended to use the most recently generated cnv file. Check print(RE) to see if cnv files from other anlyses have been generated.}

\item{splitcov}{List of small .cov dataframes. Would be stored as metadata after SARC::regionSplit.}

\item{nameofnewdf}{Name of new dataframe to be saved in metadata(RE)[['CNVlist']]. Default is CNVmeans.}
}
\value{
A new cnv file with an additional column representing the meanScores. A new list of dataframes which contains the meanscores.
}
\description{
Function to calculate MeanScores for each region where a CNV has been detected. Dependent of sequencing quality and WES/WGS input, the estimated MeanScores for CNV types are generally: HOMOZYGOUS DELETION 0-0.3, HETEROZYGOUS DELETION 0.3-0.7, HETEROZYGOUS DUPLICATION 1.3-1.8 HOMOZYGOUS DUPLICATION >1.8.
}
\examples{
data("test_cnv")
data("test_cov")
SARC <- regionSet(cnv = test_cnv, cov = test_cov)
SARC <- regionSplit(RE = SARC, cnv = metadata(SARC)[['CNVlist']][[1]],
                    startlist = metadata(SARC)[[2]],
                    endlist = metadata(SARC)[[3]])
SARC <- regionMean(RE = SARC, cnv = metadata(SARC)[['CNVlist']][[1]],
                   splitcov = metadata(SARC)[[4]])

}
