% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionQuantiles.R
\name{regionQuantiles}
\alias{regionQuantiles}
\title{regionQuantiles}
\usage{
regionQuantiles(MA, bed, meancov, q1 = 0.05, q2 = 0.95)
}
\arguments{
\item{MA}{MMultiAssayExperiment object used to store all information.}

\item{bed}{CNV bed files stored as dataframes. Chromosomes, Start of CNV, End of CNV, Type of CNV and some other information can be kept here. It is recomended to use the most recently generated bed file. Check print(MA) to see if bed files from other anlyses have been generated.}

\item{meancov}{Small cov files with meanscores included. This will be stored as metadata after being generated by SARC::regionMean.}

\item{q1}{Lower quantile cut-off. Default is .05. This should be adjusted based on the number of samples, i.e. less samples > higher the lower threshold.}

\item{q2}{Upper quantile cut-off. Default is .95. This should be adjusted based on the number of samples, i.e. less samples > lower the higher threshold.}
}
\value{
A new bed file with two new columns. For deletions the Qlow columns will be 0 or 1, and for duplications the Qhigh columns will be 0 or 1.
}
\description{
Calculates quantile distribution from each sample, within a region with a suspected CNV. We assume true deletions would be in the lower end of the quantile distribution and true duplications to be in the upper end of the quantile distribution. Many CNVs captured by detection pipelines will be false positives, this this is a good metric to remove obvious false positives. Lower quantile (q1) and upper quantile (q2) should be adjusted based on the number of patient samples available.
}
\examples{
data("test_bed")
data("test_cov")
SARC <- regionSet(bed = test_bed, cov = test_cov)
SARC <- regionSplit(MA = SARC, bed = test_bed, cov = test_cov,
                    startlist = metadata(SARC)[[1]],
                    endlist = metadata(SARC)[[2]])
SARC <- regionMean(MA = SARC, bed = test_bed, splitcov = metadata(SARC)[[3]])
SARC <- regionQuantiles(MA = SARC, bed = experiments(SARC)[[1]],
                      meancov = metadata(SARC)[[4]], q1=.15, q2=.85)
}
