% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionSet.R
\name{regionSet}
\alias{regionSet}
\title{regionSet}
\usage{
regionSet(bed, cov)
}
\arguments{
\item{bed}{CNV bed files stored as dataframes. Chromosomes, Start of CNV, End of CNV, Type of CNV and
some other information can be kept here.}

\item{cov}{cov file from WES platform/ sequencer kit or if WGS regular intervals. Stored
as a dataframe - genomic locations as rows and samples as columns.}
}
\value{
MultiAssayExperiment Object with two lists: Start sites and End sites.
}
\description{
Stores .cov and .bed file into a multiassayexperiment object. Will use genomic locations from the .bed file to find the start and end points of the CNVs from the .cov file and store these as two lists.
}
\examples{
data("test_bed")
data("test_cov")
SARC <- regionSet(bed = test_bed, cov = test_cov)
}
