% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionSplit.R
\name{regionSplit}
\alias{regionSplit}
\title{regionSplit}
\usage{
regionSplit(MA, bed, cov, startlist, endlist)
}
\arguments{
\item{MA}{MultiAssayExperiment object used to store all information.}

\item{bed}{Bed file containing CNVs which the user wishes to generate plots for. It is recommended that the most recently created bed file is used. Check print(MA) to see more bed files created by SARC.}

\item{cov}{cov file from WES platform/ sequencer kit or if WGS regular intervals. Stored as a dataframe - genomic locations as rows and samples as columns.}

\item{startlist}{List of start sites created from SARC::regionSet and stored as metadata.}

\item{endlist}{List of end sites created from SARC::regionSet and stored as metadata.}
}
\value{
A list of dataframes - each one a smaller cov file.
}
\description{
Splits the large cov file into multiple smaller cov files. Each one is specific for each CNV entry in the bam file.
}
\examples{
data("test_bed")
data("test_cov")
SARC <- regionSet(bed = test_bed, cov = test_cov)
SARC <- regionSplit(MA = SARC, bed = test_bed, cov = test_cov,
                    startlist = metadata(SARC)[[1]],
                    endlist = metadata(SARC)[[2]])
}
