% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setQDPlot.R
\name{setQDplot}
\alias{setQDplot}
\title{setQDPlot}
\usage{
setQDplot(RE, meancov)
}
\arguments{
\item{RE}{RaggedExperiment object used to store all information.}

\item{meancov}{Small cov files with mean scores included. This will be stored as metadata after being generated by SARC::regionMean.}
}
\value{
A list of dataframes. Each dataframe contains ranked meanscores for each sample. One dataframe is made for each CNV in the cnv file.
}
\description{
Quality checking plot to see the Distribution of mean scores at a given region of the genome. The sample with the suspected CNV is highlighted. Useful way of checking why a certain CNV may be a false positive.
}
\examples{
data("test_cnv")
data("test_cov")
SARC <- regionSet(cnv = test_cnv, cov = test_cov)
SARC <- regionSplit(RE = SARC, cnv = metadata(SARC)[[1]][[1]],
                    startlist = metadata(SARC)[[2]],
                    endlist = metadata(SARC)[[3]])
SARC <- regionMean(RE = SARC, cnv = metadata(SARC)[[1]][[1]],
                   splitcov = metadata(SARC)[[4]])
SARC <- setQDplot(RE = SARC, meancov = metadata(SARC)[[5]])
}
